refactor(update): split update settings UI and harden API assumptions
